+++ boosters/node-beginner/tests/01_exercise.test.js
@@ -1,19 +1,19 @@
 const request = require("supertest");
-const app = require("../src/app");
-const setup = require("./config/setup")("Users");
+const app = require("../src/server");
+const setup = require("./config/setup");
 const teardown = require("./config/teardown");
 const uuid = require("uuid/v4");
 
 const users = [
   { firstName: "Andrew", lastName: "Mayer", id: uuid() },
-  { firstName: "Andrew", lastName: "Mayer", id: uuid() },
+  { firstName: "Liang", lastName: "Chen", id: uuid() },
 ];
 
-describe("test", () => {
-  beforeAll(setup(users));
+describe("Specific user", () => {
+  beforeAll(setup("Users", users));
   afterAll(teardown);
 
-  it("should return user profile given speicifc userId", async () => {
+  it("Should return user profile given speicifc userId", async () => {
     res = await Promise.all(users.map(user => request(app).get(`/users/${user.id}`)));
 
     users.map((user, i) => {
